Non-200 status codes imply API error
Prior to this patch, the only errors
that were passed to the callback functions
were low-level request errors. Chain API
errors were not being passed as errors to the callback funciton.

When the Chain API returns a non-200 response, the body of the
response contains an error object explaining what went wrong.